% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc.acs.health-package.R
\docType{package}
\name{dc.acs.health-package}
\alias{dc.acs.health}
\alias{dc.acs.health-package}
\title{dc.acs.health: Social Data Commons: Health, Behavior, Diet: Health Insurance}
\description{
Allows user to easily get health insurance data from the SDAD Data Commons. Data can be provided directly or in a file.
}
\author{
\strong{Maintainer}: Kathryn Linehan \email{kjl5t@virginia.edu}

Authors:
\itemize{
  \item Social and Decision Analytics Division (SDAD), Biocomplexity Institute, University of Virginia
}

}
\keyword{internal}
